# feat!: GPU accelerated VID disperse