--- a/primitives/src/lib.rs
+++ b/primitives/src/lib.rs
@@ -51,9 +51,7 @@ pub mod icicle_deps {
     /// curve-specific types both from arkworks and from ICICLE
     /// including Pairing, CurveCfg, Fr, Fq etc.
     pub mod curves {
-        pub use ark_bls12_381::Bls12_381;
         pub use ark_bn254::Bn254;
-        pub use icicle_bls12_381::curve::CurveCfg as IcicleBls12_381;
         pub use icicle_bn254::curve::CurveCfg as IcicleBn254;
     }