@mrain 2024-03-14
I was flooded by profiling the advz dispersal 
`cargo test --release -p jf-primitives --features test-srs,icicle,kzg-print-trace,parallel,gpu-vid -- disperse_timer --ignored --nocapture`

It's calling `batch_commit` which launches many `commit`s in parallel, and each of them will start a timer in a nested manner b/c `kzg-print-trace` feature is on. Should we remove the inner timers in `commit`, and add cfg feature flags in `batch_commit` timers?