@mrain 2024-03-14
My local run results are listed below.
Batch commit performance is still not perfect. Will check later

```
Start:   KZG10::Setup with prover degree 256 and verifier degree 1
··Start:   Generating powers of G
··End:     Generating powers of G ..................................................3.119ms
End:     KZG10::Setup with prover degree 256 and verifier degree 1 .................4.762ms
Start:   VID disperse 33554432 payload bytes to 512 nodes
··Start:   encode payload bytes into polynomials
··End:     encode payload bytes into polynomials ...................................67.697ms
··Start:   compute all storage node evals for 4229 polynomials with 256 coefficients
··End:     compute all storage node evals for 4229 polynomials with 256 coefficients 147.929ms
··Start:   compute merkle root of all storage node evals
··End:     compute merkle root of all storage node evals ...........................74.827ms
··Start:   compute 4229 KZG commitments
··End:     compute 4229 KZG commitments ............................................138.304ms
··Start:   compute aggregate proofs for 512 storage nodes
····Start:   compute h_poly
····End:     compute h_poly ........................................................53.948ms
····Start:   gen eval proofs with parallel_factor 2 and num_points 512
····End:     gen eval proofs with parallel_factor 2 and num_points 512 .............50.675ms
··End:     compute aggregate proofs for 512 storage nodes ..........................106.559ms
··Start:   assemble shares for dispersal
··End:     assemble shares for dispersal ...........................................37.821ms
End:     VID disperse 33554432 payload bytes to 512 nodes ..........................603.895ms
Start:   VID disperse 33554432 payload bytes to 512 nodes
··Start:   encode payload bytes into polynomials
··End:     encode payload bytes into polynomials ...................................28.887ms
··Start:   compute all storage node evals for 4229 polynomials with 256 coefficients
··End:     compute all storage node evals for 4229 polynomials with 256 coefficients 56.072ms
··Start:   compute merkle root of all storage node evals
··End:     compute merkle root of all storage node evals ...........................66.781ms
··Start:   compute 4229 KZG commitments
····Start:   batch commit 4229 polynomials
····End:     batch commit 4229 polynomials .........................................1.532s
··End:     compute 4229 KZG commitments ............................................1.532s
··Start:   compute aggregate proofs for 512 storage nodes
····Start:   compute h_poly
····End:     compute h_poly ........................................................58.700ms
····Start:   gen eval proofs with parallel_factor 2 and num_points 512
····End:     gen eval proofs with parallel_factor 2 and num_points 512 .............49.792ms
··End:     compute aggregate proofs for 512 storage nodes ..........................110.447ms
··Start:   assemble shares for dispersal
··End:     assemble shares for dispersal ...........................................14.715ms
End:     VID disperse 33554432 payload bytes to 512 nodes ..........................1.836s
```